Move fix_court_names to strip_court_names.py
